Add files via upload
The system developed is a virtual board game – but it does not have an elaborate graphical user interface.  
Instead, the game is to be played via the console (e.g. Eclipse). It also has its own distinctive theme, based on NASA’s mission to land the first woman and next man on the moon by 2024 (https://nasa.gov/specials/artemis/ ).